updated speed reading on main page so that it is now an integer. changed font size and box size of this as well

+++ MainPage.xaml.cs
@@ -117,7 +117,7 @@ namespace Project
             this.Children.Add(this.finish);
         }
 
-        public void updateSpeed(float speed)
+        public void updateSpeed(int speed)
         {
             this.txtSpeed.Text = ""+ speed + " m/s"; 
         }
